+++ .storybook/webpack.config.js
@@ -1,6 +1,4 @@
-const path = require('path')
-
-module.exports = function({ config }) {
+module.exports = function ({ config }) {
   config.module.rules.push({
     test: /\.(ts|tsx)$/,
     use: [
