+++ package.json
@@ -44,6 +44,7 @@
     "react-dom": "^16.11.0",
     "react-helmet": "^5.2.0",
     "react-hook-inview": "^4.1.2",
+    "react-markdown": "^4.2.2",
     "react-share": "2.1.0",
     "react-transition-group": "^2.4.0",
     "react-window-decorators": "^1.0.2",
