+++ src/components/Jobs/index.jsx
@@ -4,16 +4,12 @@ import { PostCard } from '../PostCard'
 import ArticleHeader from '../ArticleHeader/index'
 import { data } from '../../data/jobs'
 
-export const Jobs = () => {
-  const { posts } = data
-
-  return (
-    <div>
-      <Helmet>
-        <title>Jobs @ Hardfork.io, Berlin</title>
-      </Helmet>
-      <ArticleHeader title="Jobs" subtitle="Hardfork.io | Berlin" />
-      <PostCard posts={posts} />
-    </div>
-  )
-}
+export const Jobs = () => (
+  <div>
+    <Helmet>
+      <title>Jobs @ Hardfork.io, Berlin</title>
+    </Helmet>
+    <ArticleHeader title="Jobs" subtitle="Hardfork.io | Berlin" />
+    <PostCard posts={data.posts} />
+  </div>
+)
