+++ src/data/jobs.js
@@ -1,0 +1,13 @@
+export const data = {
+  posts: [
+    {
+      title: 'Junior Blockchain Developer	',
+      slug: 'juniordev	',
+      id: 1,
+      meta_title: 'Hardfork GmbH - Job Ad Blockchain Junior Developer	',
+      meta_description: 'Hardfork is looking for a Junior Blockchain Developer	',
+      skillsDescription:
+        'We’re looking for a smart, confident and positive team player, who is willing to learn a lot and teach others about their findings. You should see the importance of writing clean code and enjoy working in a team. We expect a person with a university degree in a STEM field (Master, Diploma or above) and with a “can do anything” mentality.	In return, we promise to give you the runway to experiment, learn and grow.	',
+    },
+  ],
+}
