+++ src/data/juniorDev.md
@@ -1,0 +1,46 @@
+# Junior Blockchain Developer (f/m/i)	
+
+## The company	
+
+Hardfork is a technical consulting company in the new and rising sphere of Blockchain technology. We support our customers in choosing the right approach to solve their technical problems in a sustainable and efficient manner. Our services range from feasibility studies over prototype development to fully hosted and maintained blockchain applications (dApps).	
+
+## The job	
+
+You will be part of a small team of technical coworkers with a strong background in STEM fields, building exciting things and permanently exchanging knowledge to maximize the quality of the software we build for our clients. You will design and implement decentralized applications with full end-to-end responsibility while also shaping the way we work as a team and how modern product development should be done.	
+You’ll mainly work with these technologies:	
+
+* JavaScript (ECMA2016+, NodeJS, TypeScript)	
+* Docker	
+* Linux	
+* “Blockchain” (Ethereum, Bitcoin, potentially many others)	
+
+## Skills	
+
+We’re looking for a smart, confident and positive team player, who is willing to learn a lot and teach others about their findings. You should see the importance of writing clean code and enjoy working in a team. We expect a person with a university degree in a STEM field (Master, Diploma or above) and with a “can do anything” mentality.	
+In return, we promise to give you the runway to experiment, learn and grow.	
+
+## The play	
+
+#### 5 Hour Work Day!	
+
+At Hardfork we permanently improve our efficiency to be able to deliver in 5 hours what anybody else achieves in their 8+ hour day. Everybody here gets a full salary for working every day from 8 a.m. to 1 p.m. This process makes us much more productive and focused in our work. It forces us to permanently leave our comfort zone, innovate and improvise, so we have more time for what is most important to us besides work. We work hard to work less.	
+
+#### Hard Facts:	
+
+* Salary: 42,000 Euros per year (brutto)	
+* Work hours: 25 per week (8 a.m. to 1 p.m., 5 days a week, no exceptions, no Gleitzeit - sorry)	
+* Vacation: 6 weeks per year paid vacation	
+
+## The process	
+
+If you are up for an exciting experiment in work life balance and pushing the frontier of technology, apply now!	
+Please send an email to jobs@hardfork.io and include:	
+
+* A compact CV,	
+* A copy of your thesis and graduation certificate including grades	
+* A short text about what you expect for your personal and professional life from the five hour work day,	
+
+We will require later:	
+
+* References (contacts of persons you worked with, e.g. professors, employers etc. we will ask you before we contact any of them), please no “reference letters”,	
+* Link to profile on: Github, Stackexchange, etc.
