+++ typings/css.d.ts
@@ -1,1 +1,1 @@
-declare module "*.scss"
+declare module '*.scss'
